Add macOS rebuild memory budget backend
Share memory availability detection between rebuild selection and transformer loading. Keep the existing cgroup/procfs Linux behavior and add a Mach fallback so macOS can create and enforce live rebuild budgets.
