+++ libtenzir/include/tenzir/detail/available_memory.hpp
@@ -1,0 +1,24 @@
+//
+//  ▀▀█▀▀ █▀▀▀ █▄  █ ▀▀▀█▀ ▀█▀ █▀▀▄
+//    █   █▀▀  █ ▀▄█  ▄▀    █  █▀▀▄
+//    ▀   ▀▀▀▀ ▀   ▀ ▀▀▀▀▀ ▀▀▀ ▀  ▀
+//
+// SPDX-FileCopyrightText: (c) 2026 The Tenzir Contributors
+// SPDX-License-Identifier: BSD-3-Clause
+
+#pragma once
+
+#include <cstdint>
+#include <optional>
+#include <string>
+
+namespace tenzir::detail {
+
+struct available_memory_info {
+  uint64_t bytes = 0;
+  std::string source = {};
+};
+
+auto available_memory() -> std::optional<available_memory_info>;
+
+} // namespace tenzir::detail
